clean collection meta, move minmax out
